+++ Bridge/Resources/StringsResourceManager.cs
@@ -1,3 +1,4 @@
+using System.Globalization;
 using System.Resources;
 
 namespace Bridge.Resources;
@@ -8,5 +9,7 @@ internal static class StringsResourceManager
         "Bridge.Resources.Strings",
         typeof(StringsResourceManager).Assembly);
 
-    public static string? GetString(string name) => Manager.GetString(name);
+    public static CultureInfo Culture { get; set; } = CultureInfo.GetCultureInfo("en");
+
+    public static string? GetString(string name) => Manager.GetString(name, Culture);
 }